+++ config/workdays.php
@@ -18,6 +18,12 @@ return [
         'adjustment' => 0,
     ],
 
+    'carbon_macros' => [
+        'enabled' => true,
+        'short_aliases' => true,
+        'override_existing' => false,
+    ],
+
     'profiles' => [
         'iran' => [
             'weekends' => ['Thursday', 'Friday'],